Refactored admin collection
