+++ frontend/public/collection.css
@@ -2,7 +2,7 @@
 .collection-page-container {
   height: 100vh;
   width: 100%;
-  background-image: url("ocean.jpg");
+  background-image: url("mini_tree.jpg");
   background-size: cover;
   overflow: scroll;
   position: relative;
