+++ frontend/public/index.html
@@ -27,11 +27,10 @@
     -->
     <title>KRMA</title>
     <link rel="stylesheet" type="text/css" href="reset.css">
-    <link rel="stylesheet" type="text/css" href="/styling/app.css">
     <link rel="stylesheet" type="text/css" href="splash.css">
     <link rel="stylesheet" type="text/css" href="session_forms.css">
     <link rel="stylesheet" type="text/css" href="navbar.css">
-    <link rel="stylesheet" type="text/css" href="treasureisland.css">
+    <link rel="stylesheet" type="text/css" href="main_page.css">
     <link rel="stylesheet" type="text/css" href="collection.css">
     <link rel="stylesheet" type="text/css" href="add_karma.css">
     <link rel="stylesheet" type="text/css" href="report.css">
