+++ frontend/public/main_page.css
