+++ frontend/public/usersPage.css
@@ -31,7 +31,7 @@
   justify-content: space-between;
   align-items: center;
   width: 650px;
-  border: solid 1px white;
+  border: solid 1px #797976;
   border-radius: 5px;
   background-color: #213637;
   color: white;
