+++ frontend/src/components/adminbar/admin_bar.jsx
@@ -13,8 +13,8 @@ class AdminBar extends React.PureComponent {
         <div className="navbar-right">
           <div className="navbar-items">
             <Link to='/users'>All Users</Link>
-            <Link to='/contents'>All Content</Link>
-            <Link to='/reports'>Reported Content</Link>
+            <Link to='/collection'>All Karma</Link>
+            <Link to='/reports'>Reported Karma</Link>
             <div onClick={() => logout()}>Logout</div>
           </div>
         </div>
