+++ frontend/src/components/app.jsx
@@ -7,7 +7,7 @@ import SignupFormContainer from './session/signup_form_container';
 import LoginFormContainer from './session/login_form_container';
 import SplashPageContainer from './splash_page/splash_page_container';
 import ReportsContainer from './moderator/report_inbox/reports_container';
-import TreasureContentsContainer from './moderator/box_contents/contents_container';
+// import TreasureContentsContainer from './moderator/box_contents/contents_container';
 import UsersContainer from './moderator/all_users/users_container';
 import Modal from './modal/modal';
 
@@ -21,7 +21,7 @@ const App = () => (
       <ProtectedRoute exact path="/main" component={TreasureIslandContainer} />
       <ProtectedRoute exact path="/collection" component={CollectionContainer} />
       <AdminRoute exact path="/reports" component={ReportsContainer} />
-      <AdminRoute exact path="/contents" component={TreasureContentsContainer} />
+      {/* <AdminRoute exact path="/contents" component={TreasureContentsContainer} /> */}
       <AdminRoute exact path="/users" component={UsersContainer} />
       <Redirect path="/" to="/" />
     </Switch>
