+++ frontend/src/components/moderator/box_contents/admin_item.js
@@ -1,3 +1,5 @@
+// TBD
+
 import React from 'react';
 import Modal from '../../modal/modal';
