+++ frontend/src/components/moderator/box_contents/admin_item_container.js
@@ -1,3 +1,5 @@
+// TBD
+
 import { connect } from 'react-redux';
 import AdminContentItem from './admin_item';
 import { deleteTreasure } from '../../../actions/treasure_actions';
