+++ frontend/src/components/moderator/box_contents/contents.jsx
@@ -1,7 +1,8 @@
+// TBD
+
 import React from 'react';
 import AdminItemContainer from './admin_item_container';
 import AdminNavBar from '../../adminbar/admin_bar_container';
-import NavBarContainer from '../../navbar/navbar_container';
 
 class TreasureContentsPage extends React.Component {
   componentDidMount() {
@@ -14,8 +15,7 @@ class TreasureContentsPage extends React.Component {
     return (
       <div className="contents-page">
         <AdminNavBar />
-        <NavBarContainer />
-        <h1 className="contents-header">Everything Inside the Treasure Box</h1>
+        <h1 className="contents-header">All Karma</h1>
         <div className="the-collection">
           {allTreasures.map((treasure) => (
             <AdminItemContainer key={treasure._id} treasure={treasure} />
