+++ frontend/src/components/moderator/box_contents/contents_container.js
@@ -1,3 +1,5 @@
+// TBD
+
 import { connect } from 'react-redux';
 import { fetchTreasures } from '../../../actions/treasure_actions';
 import TreasureContentsPage from './contents';
